+++ app/models/break.rb
@@ -5,9 +5,10 @@ class Break < ApplicationRecord
   after_destroy :refresh_teaching_period_communication_schedule_caches
 
   validates :start_date, presence: true
-  validates :number_of_weeks, presence: true
+  validates :number_of_days, presence: true
   validates :teaching_period_id, presence: true
   validate :ensure_campuses_exist
+  validate :ensure_pause_week_count_is_whole_weeks
 
   validate :ensure_start_date_is_within_teaching_period, :ensure_break_end_is_within_teaching_period, :ensure_break_is_not_colliding
 
@@ -18,8 +19,8 @@ class Break < ApplicationRecord
   end
 
   def ensure_break_end_is_within_teaching_period
-    if start_date + number_of_weeks.weeks > teaching_period.end_date
-      errors.add(:number_of_weeks, "is exceeding Teaching Period end date")
+    if start_date + number_of_days.days > teaching_period.end_date
+      errors.add(:number_of_days, "is exceeding Teaching Period end date")
     end
   end
 
@@ -35,7 +36,13 @@ class Break < ApplicationRecord
   end
 
   def duration
-    number_of_weeks.weeks
+    number_of_days.days
+  end
+
+  # The number of whole teaching weeks this break spans - partial weeks count as
+  # a full week, as teaching weeks cannot be split.
+  def weeks_spanned
+    (number_of_days / 7.0).ceil
   end
 
   def first_monday
@@ -46,7 +53,7 @@ class Break < ApplicationRecord
   end
 
   def monday_after_break
-    first_monday + number_of_weeks.weeks
+    first_monday + number_of_days.days
   end
 
   def end_date
@@ -70,6 +77,15 @@ class Break < ApplicationRecord
     campus_ids.blank? ||
       other_break.campus_ids.blank? ||
       campus_ids.map(&:to_i).intersect?(other_break.campus_ids.map(&:to_i))
+  end
+
+  def ensure_pause_week_count_is_whole_weeks
+    return unless pause_week_count?
+    return if number_of_days.blank?
+
+    unless (number_of_days % 7).zero?
+      errors.add(:pause_week_count, 'can only be set on breaks that are a multiple of 7 days')
+    end
   end
 
   def ensure_campuses_exist