+++ app/models/comments/assessment_comment.rb
@@ -1,6 +1,7 @@
 class AssessmentComment < TaskComment
   before_create do
     self.content_type = :assessment
+    self.attention_audience = :student
   end
 
   def serialize(user)