+++ db/migrate/20260820054718_add_group_references_to_communication_conditions.rb
@@ -1,0 +1,6 @@
+class AddGroupReferencesToCommunicationConditions < ActiveRecord::Migration[8.0]
+  def change
+    add_reference :communication_conditions, :group_set
+    add_reference :communication_conditions, :group
+  end
+end